Crate opentaws

This is a proof of concept TAWS as described in DO-367. It is not even close to fulfilling DO-367 C, the simplest TAWS class. It exists to learn about using BDD (Cucumber & Gherkin in particular) for implementing avionic.

Using openTAWS

Currently it is only possible to use openTAWS from Rust. We’ve looked briefly into WASM-WASI and C ABI as addiotional targets, but this did not lead anywehre usable so far. We are very open to suggestions, so please open an issue if you have some feedback.

Modules

functionalities

prelude

The prelude is a collection of all traits and commonly used types in this crate

Structs

AircraftState

Represents the current state of an aircraft

AlertState

Collection of a all alerts which are currently present in the TAWS

Taws

Represents one instance of a TAWS

TawsConfig

This configuration holds various details about the aircraft in use. These are necessary for example when estimating path trajectories for FLTA.

Enums

Alert

Available alerts from the TAWS.

AlertLevel

Importance level of an alert