Module opentaws::prelude

The prelude is a collection of all traits and commonly used types in this crate

For normal use of this crate it is sufficient to glob import only this moduel, e.g. use opentaws::prelude::*.

Re-exports

pub use crate::Taws;

Structs

AircraftState

Represents the current state of an aircraft

AlertState

Collection of a all alerts which are currently present in the TAWS

TawsConfig

This configuration holds various details about the aircraft in use. These are necessary for example when estimating path trajectories for FLTA.

degree

degrees

foot

feet

foot_per_minute

feet per minute

foot_per_second_squared

feet per second squared

knot

knots

second

The second is the SI unit of time. It is defined by taking the fixed numerical value of the caesium frequency ∆νCs, the unperturbed ground-state hyperfine transition frequency of the caesium 133 atom, to be 9 192 631 770 when expressed in the unit Hz, which is equal to s⁻¹.

Enums

Alert

Available alerts from the TAWS.

AlertLevel

Importance level of an alert

Traits

AlertSystem

Trait which is to be fulfilled by all functionalities

Type Definitions

Acceleration

Quantity type alias using the default base units.

AmountOfSubstance

Quantity type alias using the default base units.

Angle

Quantity type alias using the default base units.

AngularAcceleration

Quantity type alias using the default base units.

AngularJerk

Quantity type alias using the default base units.

AngularVelocity

Quantity type alias using the default base units.

Area

Quantity type alias using the default base units.

AvailableEnergy

Quantity type alias using the default base units.

Capacitance

Quantity type alias using the default base units.

CatalyticActivity

Quantity type alias using the default base units.

CatalyticActivityConcentration

Quantity type alias using the default base units.

Curvature

Quantity type alias using the default base units.

ElectricCharge

Quantity type alias using the default base units.

ElectricCurrent

Quantity type alias using the default base units.

ElectricPotential

Quantity type alias using the default base units.

ElectricalConductance

Quantity type alias using the default base units.

ElectricalResistance

Quantity type alias using the default base units.

Energy

Quantity type alias using the default base units.

Force

Quantity type alias using the default base units.

Frequency

Quantity type alias using the default base units.

HeatCapacity

Quantity type alias using the default base units.

HeatFluxDensity

Quantity type alias using the default base units.

HeatTransfer

Quantity type alias using the default base units.

Inductance

Quantity type alias using the default base units.

Information

Quantity type alias using the default base units.

InformationRate

Quantity type alias using the default base units.

Jerk

Quantity type alias using the default base units.

Length

Quantity type alias using the default base units.

Luminance

Quantity type alias using the default base units.

LuminousIntensity

Quantity type alias using the default base units.

MagneticFlux

Quantity type alias using the default base units.

MagneticFluxDensity

Quantity type alias using the default base units.

Mass

Quantity type alias using the default base units.

MassConcentration

Quantity type alias using the default base units.

MassDensity

Quantity type alias using the default base units.

MassRate

Quantity type alias using the default base units.

MolarConcentration

Quantity type alias using the default base units.

MolarEnergy

Quantity type alias using the default base units.

MolarMass

Quantity type alias using the default base units.

Momentum

Quantity type alias using the default base units.

Power

Quantity type alias using the default base units.

Pressure

Quantity type alias using the default base units.

RadiantExposure

Quantity type alias using the default base units.

Ratio

Quantity type alias using the default base units.

SolidAngle

Quantity type alias using the default base units.

SpecificHeatCapacity

Quantity type alias using the default base units.

TemperatureInterval

Quantity type alias using the default base units.

ThermalConductivity

Quantity type alias using the default base units.

ThermodynamicTemperature

Quantity type alias using the default base units.

Time

Quantity type alias using the default base units.

Torque

Quantity type alias using the default base units.

V

Storage type.

Velocity

Quantity type alias using the default base units.

Volume

Quantity type alias using the default base units.

VolumeRate

Quantity type alias using the default base units.