Trait opentaws::prelude::AlertSystem

pub trait AlertSystem: Debug + Send {
    // Required methods
    fn new(config: &TawsConfig) -> Self
       where Self: Sized;
    fn is_armed(&self) -> bool;
    fn arm(&mut self);
    fn disarm(&mut self);
    fn inhibit(&mut self);
    fn uninhibit(&mut self);
    fn is_inhibited(&self) -> bool;
    fn process(&mut self, state: &AircraftState) -> Option<AlertLevel>;
}

Trait which is to be fulfilled by all functionalities

Required Methods

fn new(config: &TawsConfig) -> Selfwhere Self: Sized,

Allows this system to be instantiated

fn is_armed(&self) -> bool

Returns whether this alarm is armed.

Arming refers to the automatic switching on of a function by the Equipment (DO-367 Chapter 1.9).

fn arm(&mut self)

Arm this alert

fn disarm(&mut self)

Disarm this alert

fn inhibit(&mut self)

Dismiss this alert

fn uninhibit(&mut self)

Enable this alert

fn is_inhibited(&self) -> bool

Returns whether this alarm is inhibited

fn process(&mut self, state: &AircraftState) -> Option<AlertLevel>

Process a new AircraftState, emit alerts if appropiate